feat(comments): add delete comment api

+++ internal/app/driven/commentsRepo/sql/queries.sql
@@ -32,4 +32,5 @@ WHERE
 -- name: delete :exec
 DELETE FROM comments
 WHERE
-  id = ?;
+  id = ?
+  AND author_id = ?;
